refactor: recruitment banner
